move renderer to core subpackage
