+++ input_test.go
@@ -48,7 +48,7 @@ func TestInputRender(t *testing.T) {
 	for _, test := range tests {
 		outputBuffer.Reset()
 		test.data.Input = test.prompt
-		err := test.prompt.render(
+		err := test.prompt.Render(
 			InputQuestionTemplate,
 			test.data,
 		)
